+++ CHANGELOG.md
@@ -30,6 +30,10 @@ To install release candidates run `[sudo] gem install cocoapods --pre`
 
 ##### Bug Fixes
 
+* Warn when a pod that was added or changed includes script phases  
+  [Dimitris Koutsogiorgas](https://github.com/dnkoutso)
+  [#7110](https://github.com/CocoaPods/CocoaPods/pull/7110)
+
 * Build pod targets with script phases and integrate them properly  
   [Dimitris Koutsogiorgas](https://github.com/dnkoutso)
   [#7104](https://github.com/CocoaPods/CocoaPods/pull/7104)